Fixed issues with legacy C++ compilers

+++ include/sexp/sexp-error.h
@@ -60,6 +60,6 @@ class sexp_exception_t : public std::exception {
 };
 
 void sexp_error(
-  sexp_exception_t::severity level, const char *msg, uint32_t c1, uint32_t c2, int pos);
+  sexp_exception_t::severity level, const char *msg, size_t c1, size_t c2, int pos);
 
 } // namespace sexp
