+++ src/sexp-error.cpp
@@ -45,7 +45,7 @@ std::string sexp_exception_t::format(std::string message, severity level, int po
 };
 
 void sexp_error(
-  sexp_exception_t::severity level, const char *msg, uint32_t c1, uint32_t c2, int pos)
+  sexp_exception_t::severity level, const char *msg, size_t c1, size_t c2, int pos)
 {
     char                       tmp[256];
     sexp_exception_t::severity l = (sexp_exception_t::severity) level;
